+++ lib/utils/clone_manager.dart
@@ -10,6 +10,7 @@ import 'package:clonify/src/clonify_core.dart';
 import 'package:clonify/utils/asset_manager.dart';
 import 'package:clonify/utils/background_geolocation_license_manager.dart';
 import 'package:clonify/utils/clonify_helpers.dart' hide saveLastClientId;
+import 'package:clonify/utils/notification_icon_manager.dart';
 import 'package:clonify/utils/firebase_manager.dart';
 import 'package:clonify/utils/package_rename_plus_manager.dart';
 import 'package:clonify/utils/shorebird_manager.dart';
@@ -933,6 +934,7 @@ Future<Map<String, dynamic>?> configureApp(
 
     generateCloneConfigFile(CloneConfigModel.fromJson(configJson));
     await applyBackgroundGeolocationLicenses(configJson);
+    await applyAndroidNotificationIcon(callModel.clientId!, configJson);
 
     logger.i('✅ Successfully cloned app for ${callModel.clientId}!');
     return configJson;